Add submodule support to App sandbox constructor.
This commit makes it possible to use submodules via the sandbox constructor
pattern. This is helpful when the need arises for better code organization.

For example, right now we have the `App.modules.Map` module. If we wanted to, we
could now add a submodule to it like this: `App.modules.Map.Layers`. To make this
work, modules just need to declare submodules by name:

```javascript
// map.js
App.modules.Map = function(app) {
  app.modules.Map.submodules = ['Layers'];
  // ...
};

// map.layers.js
App.modules.Map.Layers = function(app) {
  // ...
};
```

Note that only a single level of submodules are supported. So we can do this:

*  `App.modules.Map.Layers`
*  `App.modules.Map.Controls`
*  `App.modules.Map.Events`

But not this:

*  `App.modules.Map.Layers.Controls`
*  `App.modules.Map.Controls.Displays`
*  `App.modules.Map.Events.UI`